Converted most GoG.Infrastructure classes to C++ and moved into Fuego so that more code can be located there.

+++ GoG.WinRT/Common/ListCollectionView.cs
@@ -471,7 +471,7 @@ namespace GoG.WinRT.Common
         class VectorChangedEventArgs : IVectorChangedEventArgs
         {
             readonly CollectionChange _cc = CollectionChange.Reset;
-            readonly uint _index = (uint)0xffff;
+            readonly uint _index = 0xffff;
 
             static readonly VectorChangedEventArgs _reset = new VectorChangedEventArgs(CollectionChange.Reset);
             public static VectorChangedEventArgs Reset
